+++ src/img/lib/imgadm.js
@@ -78,6 +78,9 @@ var VMADM_IMG_NAME_RE = /^([a-zA-Z][a-zA-Z\._-]*)\/([0-9a-f]{8}-[0-9a-f]{4}-[0-9
 var UUID_RE = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$/;
 /* END JSSTYLED */
 
+var UA = 'imgadm/' + common.getVersion()
+    + ' (' + 'node/' + process.versions.node + '; '
+    + 'OpenSSL/' + process.versions.openssl + ')';
 
 
 // ---- internal support stuff
@@ -88,6 +91,20 @@ function _indent(s, indent) {
     return indent + lines.join('\n' + indent);
 }
 
+
+function getSysinfo(log, callback) {
+    assert.object(log, 'log');
+    assert.func(callback, 'callback');
+    exec('sysinfo', function (err, stdout, stderr) {
+        if (err) {
+            callback(err);
+        } else {
+            // Explicitly want to abort/coredump on this not being parsable.
+            var sysinfo = JSON.parse(stdout.trim());
+            callback(null, sysinfo);
+        }
+    });
+}
 
 /**
  * Call `zfs destroy -r` on the given dataset name.
@@ -704,14 +721,16 @@ IMGADM.prototype.clientFromSource = function clientFromSource(
             agent: false,
             url: baseNormUrl,
             log: self.log.child({component: 'api', source: source.url}, true),
-            rejectUnauthorized: (process.env.IMGADM_INSECURE !== '1')
+            rejectUnauthorized: (process.env.IMGADM_INSECURE !== '1'),
+            userAgent: UA
         });
     } else {
         self._clientCache[normUrl] = imgapi.createClient({
             agent: false,
             url: normUrl,
             log: self.log.child({component: 'api', source: source.url}, true),
-            rejectUnauthorized: (process.env.IMGADM_INSECURE !== '1')
+            rejectUnauthorized: (process.env.IMGADM_INSECURE !== '1'),
+            userAgent: UA
         });
     }
     callback(null, self._clientCache[normUrl]);
@@ -2025,6 +2044,7 @@ IMGADM.prototype.createImage = function createImage(options, callback) {
     var prepareTimeout = options.prepareTimeout || 300;  // in seconds
 
     var vmInfo;
+    var sysinfo;
     var vmZfsFilesystemName;
     var vmZfsSnapnames;
     var originInfo;
@@ -2082,6 +2102,17 @@ IMGADM.prototype.createImage = function createImage(options, callback) {
                 next();
             });
         },
+        function getSystemInfo(next) {
+            if (vmInfo.brand === 'kvm') {
+                next();
+                return;
+            }
+            // We need `sysinfo` for smartos images. See below.
+            getSysinfo(log, function (err, sysinfo_) {
+                sysinfo = sysinfo_;
+                next(err);
+            });
+        },
         function gatherManifest(next) {
             var m = {
                 v: common.MANIFEST_V,
@@ -2099,6 +2130,7 @@ IMGADM.prototype.createImage = function createImage(options, callback) {
                     'users', 'billing_tags', 'traits', 'generate_passwords',
                     'inherited_directories', 'nic_driver', 'disk_driver',
                     'cpu_type', 'image_size'];
+                // TODO Should this *merge* requirements?
                 INHERITED_FIELDS.forEach(function (field) {
                     if (!m.hasOwnProperty(field)
                         && originManifest.hasOwnProperty(field))
@@ -2111,6 +2143,21 @@ IMGADM.prototype.createImage = function createImage(options, callback) {
                         }
                     }
                 });
+            }
+            if (vmInfo.brand !== 'kvm' /* i.e. this is a smartos image */
+                && !(options.manifest.requirements
+                    && options.manifest.requirements.min_platform))
+            {
+                // Unless an explicit min_platform is provided (possibly empty)
+                // the min_platform for a SmartOS image must be the current
+                // platform, b/c that's the SmartOS binary compat story.
+                if (!m.requirements)
+                    m.requirements = {};
+                m.requirements.min_platform = {};
+                m.requirements.min_platform[sysinfo['SDC Version']]
+                    = sysinfo['Live Image'];
+                log.debug({min_platform: m.requirements.min_platform},
+                    'set smartos image min_platform to current');
             }
             if (incremental) {
                 if (!originInfo) {
@@ -2272,7 +2319,8 @@ IMGADM.prototype.createImage = function createImage(options, callback) {
                 key: 'prepare-image:state',
                 // Don't explicitly check for value=running here because it is
                 // fine if it blows by to 'success' between our polling.
-                timeout: prepareTimeout * 1000
+                timeout: prepareTimeout * 1000,
+                interval: 2000
             };
             log.debug('wait for up to %ds for prepare-image:state signal '
                 + 'from operator-script', prepareTimeout);
@@ -2292,8 +2340,14 @@ IMGADM.prototype.createImage = function createImage(options, callback) {
                          *   'running'
                          * - the prepare-image script crashed early
                          */
+                        logCb('Timeout waiting for prepare-image script to '
+                            + 'signal it started');
+                        log.debug('timeout waiting for operator-script to '
+                            + 'set prepare-image:state');
                         next(new errors.PrepareImageDidNotRunError(vmUuid));
                     } else {
+                        log.debug(err, 'unexpected error waiting for '
+                            + 'operator-script to set prepare-image:state');
                         next(err);
                     }
                     return;
@@ -2683,7 +2737,9 @@ IMGADM.prototype.publishImage = function publishImage(opts, callback) {
     var client = imgapi.createClient({
         agent: false,
         url: opts.url,
-        log: self.log.child({component: 'api', url: opts.url}, true)
+        log: self.log.child({component: 'api', url: opts.url}, true),
+        rejectUnauthorized: (process.env.IMGADM_INSECURE !== '1'),
+        userAgent: UA
     });
     var uuid = manifest.uuid;
     var rollbackImage;